\name{myTst2-package}
\alias{myTst2-package}
\alias{myTst2}
\docType{package}
\title{
\packageTitle{myTst2}
}
\description{
\packageDescription{myTst2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{myTst2}
\packageIndices{myTst2}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{myTst2}

Maintainer: \packageMaintainer{myTst2}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
